Version 1.92
Retested serial_reliable_datagram_client.pde and
serial_reliable_datagram_server.pde built on Linux as described in their
headers, and with USB-RS485 adapters. No changes, working correctly.
Testing of nRF5232 with Sparkfun nRF52 board support 0.2.3 shows that
there appears to be a problem with interrupt handlers on this board, and
none of the interrupt based radio drivers can be expected to work with
this chip. Ensured all interrupt routines are flagged with
ICACHE_RAM_ATTR when compiled for ESP8266, to prevent crashes.