Version 1.94
Fixed a bug in RHSoftwareSPI where RHGenericSPI::setBitOrder() has no
effect for on RHSoftwareSPI. Reported by Peter.
Added support in RHRouter for a node to optionally be leaf node, and not
participate as a router in the network. See RHRouter::setNodeTypePatch
from Alex Evans.
Fixed a problem with ESP32 causing compile errors over missing
SPI.usingInterrupt().