Adds RH_PLATFORM_PIDUINO
PiDuino is a C ++ library for Pi boards that allows the use of I/O like GPIO, I2C, SPI, UART ... with an API as close as possible to the Arduino language.
The description of Pi cards uses a stored "Object" model in a database that allows to add new models of boards easily.

At this time, the SoC models supported are AllWinner H-Series and Broadcom BCM2708 through 2710 which allows it to be used on Raspberry Pi and most Nano Pi, Orange Pi and Banana Pi.

https://github.com/epsilonrt/piduino